Fixed my tasks and UI again again
