+++ app/src/main/AndroidManifest.xml
@@ -9,24 +9,22 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ZoneInteger.IntegerAct2Activity">
-
-        </activity>
         <activity android:name=".ZoneModulo.NumberActivity">
 
         </activity>
         <activity android:name=".Result" />
-        <activity android:name=".ZoneInteger.CompositeActivity2" />
-        <activity android:name=".ZoneInteger.IntegerAct4Activity" >
+        <activity android:name=".ZoneInteger.CompositeActivity2" >
+
+        </activity>
+        <activity android:name=".ZoneInteger.Multiple2Activity" >
+
+        </activity>
+        <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
 
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
-        </activity>
-        <activity android:name=".ZoneInteger.Multiple2Activity" />
-        <activity android:name=".MainActivity">
-
         </activity>
     </application>
