+++ app/src/main/java/com/makentoshe/mathworks/MainActivity.kt
@@ -8,7 +8,7 @@ class MainActivity : AppCompatActivity() {
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
+        setContentView(R.layout.example_main)
 
     }
 }
