--- app/src/main/res/layout/example_theory.xml
@@ -1,117 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent">
-
-    <TextView
-        android:id="@+id/head"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="8dp"
-        android:text="@string/Head"
-        android:textAllCaps="true"
-        android:textColor="@color/colorHeads"
-        android:textSize="20sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.498"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent" />
-
-    <TextView
-        android:id="@+id/subhead"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="5dp"
-        android:fontFamily="sans-serif"
-        android:text="@string/Subhead1_4"
-        android:textColor="@color/colorSubheads"
-        android:textSize="15sp"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/head" />
-
-    <ProgressBar
-        android:id="@+id/progressBarTheory"
-        style="?android:attr/progressBarStyleHorizontal"
-        android:layout_width="260dp"
-        android:layout_height="18dp"
-        android:layout_marginTop="5dp"
-        android:max="8"
-        android:progress="0"
-        android:progressTint="@color/colorRedProgress"
-        android:secondaryProgressTint="@color/colorSubheads"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/subhead" />
-
-    <ProgressBar
-        android:id="@+id/progressBarTheoryTrue"
-        style="?android:attr/progressBarStyleHorizontal"
-        android:layout_width="260dp"
-        android:layout_height="18dp"
-        android:layout_marginTop="5dp"
-        android:max="8"
-        android:progress="0"
-        android:secondaryProgressTint="@color/colorSubheads"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/subhead" />
-
-    <ScrollView
-        android:id="@+id/theoryScroll"
-        android:layout_width="0dp"
-        android:layout_height="match_parent"
-        android:layout_marginTop="100dp"
-        android:layout_marginBottom="50dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent">
-
-        <LinearLayout
-            android:id="@+id/theoryLinear"
-            android:layout_width="match_parent"
-            android:layout_height="wrap_content"
-            android:orientation="vertical">
-
-            <TextView
-                android:id="@+id/theoryText"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:paddingLeft="10sp"
-                android:paddingRight="10sp"
-                android:text="@string/textMultiple1"
-                android:textColor="@color/colorText"
-                android:textSize="16sp" />
-
-            <ImageView
-                android:id="@+id/theoryImage"
-                android:layout_width="match_parent"
-                android:layout_height="match_parent"
-                android:layout_marginTop="-50dp"
-                android:contentDescription="@string/theoryImage"
-                android:scaleX="0.9"
-                android:scaleY="0.9"
-                android:visibility="gone"
-                app:srcCompat="@drawable/table_multiple" />
-
-        </LinearLayout>
-    </ScrollView>
-
-    <Button
-        android:id="@+id/continueButtonTheory"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        android:background="@color/colorBackButton"
-        android:paddingLeft="10sp"
-        android:paddingRight="10sp"
-        android:text="@string/Continue"
-        android:textColor="#FFFFFF"
-        android:textSize="25sp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
