+++ app/src/main/res/values/strings.xml
@@ -17,7 +17,7 @@
     <string name="Head1">Целые числа</string>
     <string name="Subhead1_1">Умножение и деление</string>
     <string name="Head1_4">Целые числа</string>
-    <string name="Subhead1_4">Умножение и деление</string>
+    <string name="Subhead1_4">Простые числа</string>
     <string name="Subhead">Название акта</string>
     <string name="textMultiple1">"Умножение целых чисел есть такое действие, в котором нужно взять одно число слагаемым столько раз, сколько в другом содержится единиц, и найти сумму этих слагаемых.\n Умножить 7 на 3 значит взять число 7 слагаемым три раза и найти сумму. Искомая сумма есть 21.\n <b>7 · 3 = 7 + 7 + 7 = 21</b> \n  Также при умножении, как и при сложении выполняется правило: \n <b> Произведение не изменяется от перемены порядка множителей.</b> \n  <b>7 · 3 = 3 · 7 = 3 + 3 + 3 + 3 + 3 + 3 + 3 = 21</b> \n Так как умножение многозначных чисел построено на умножении однозначных, была создана таблица умножения. С её помошью, можно найти любое произведение однозначных чисел. Например, чтобы найти произведение 5 · 3, нужно найти столбец с 5 и найти пересечение со строй с 3. Число на пересечении и есть произведение."</string>
     <string name="textMultiple2"> "Деление есть такое действие, в котором по двум данным числам определяют, сколько раз одно число содержится в другом. Данные числа в делении называются делимым и делителем, искомое называется частным. \n Делимое есть то число, которое содержит другое.Делитель есть то число, которое содержится в другом.Частное показывает, сколько раз делитель содержится в делимом. \n В данном примере делимое есть 6, делитель 2, частное 3. Разделить 6 на 2 значит также разбить 6 на 2 равных слагаемых и отыскать их величину. \n Число 6 представится при помощи двух равных слагаемых в виде: \n <b>6 = 3 + 3</b> \n <b>Способы деления:</b> \n Имея два числа 12 и 4, мы можем разделить 12 на 4 различными способами. \n С помощью сложения мы можем определить, сколько раз нужно взять 4 слагаемым для того, чтобы получить в сумме 12. Так, взяв 4 слагаемым 3 раза, находим в сумме:\n 4 + 4 + 4 = 12, \n следовательно, 4 содержится в 12 три раза. \n С помощью вычитания определяем, сколько раз можно из большего числа 12 вычесть меньшее 4. При этом мы вычитаем делитель до тех пор, пока это возможно. Так, вычитая последовательно из 12 по 4, имеем:\n 12 - 4 = 8\n 8 - 4 = 4\n 4 - 4 = 0 \n Отсюда находим, что можно вычесть 4 из 12 ровно три раза.Деление есть сокращенное вычитание равных вычитаемых.Наконец, посредством умножения, мы можем определить, на какое число нужно помножить 4, чтобы получить 12. Умножая последовательно 4 на 1, 2, 3, находим, что для того, чтобы получить 12, нужно 4 помножить на 3."</string>
@@ -27,13 +27,13 @@
     <string name="MultipleTask4">Покупатель истратил на первую покупку %1$d рублей, а на вторую - в %2$d больше. Сколько денег он истратил на обе покупки?</string>
     <string name="DivTask">Разделите %1$d на %2$d и напишите ответ</string>
     <string name="DivTask2">Всего %1$d коробок. В каждую коробку положили одинаковое количество кубиков. Всего %2$d кубиков. Сколько кубиков в каждой коробке?</string>
-    <string name="DivTask3">Пароход перевёз %1$d автомобилей, а второй - в %2$d меньше. Сколько всего автомобилей было привезено?</string>
-    <string name="CompositeTask">Разложи число %1$d на простые множители и запиши снизу подряд без пробелов в возрастающем порядке, не учитывая единицу( например для числа 6: 23)</string>
-    <string name="CompositeTask2">В книге рекордов Гиннесса написано, что наибольшее известное простое число равно  %1$d1 ^ %2$d – 1.  Напиши последнюю цифру данной разности. И напиши без пробелов и запятых после этого 0, если разность является простым числом, и 1, если разность является составным числом</string>
+    <string name="DivTask3">Пароход привёз %1$d автомобилей, а второй - в %2$d меньше. Сколько всего автомобилей было привезено?</string>
+    <string name="CompositeTask">Разложите число %1$d на простые множители и запишите снизу подряд без пробелов в возрастающем порядке, не учитывая единицу( например для числа 6: 23)</string>
+    <string name="CompositeTask2">В книге рекордов Гиннесса написано, что наибольшее известное простое число равно  %1$d1 ^ %2$d – 1. Напишите без пробелов и запятых после этого 0, если разность является простым числом, и 1, если разность является составным числом</string>
     <string name="CompositeTask3">Найдите все такие натуральные числа p, что p и  5p + 1  – простые.</string>
-    <string name="CompositeTask4">Найти все такие натуральные числа p, что p и  3p² + 1  – простые.</string>
-    <string name="NumberTask">Сложи %1$s на %2$s в %3$s -чной системе счисления и напиши ответ в этой же системе счисления </string>
+    <string name="CompositeTask4">Найдите все такие натуральные числа p, что p и  3p² + 1  – простые.</string>
+    <string name="NumberTask">Сложите %1$s на %2$s в %3$s -чной системе счисления и напишите ответ в этой же системе счисления </string>
     <string name="Number2Task">Саша живёт в стране, где у всех по %1$s пальцев а руках и соответственно имеют такую систему счисления. Помогите ему сосчитать сколько будет %2$s - %3$s. Запишите ответ в той же системе счисления.</string>
-    <string name="Number3Task">В %1$s-чном мире случилась беда. Зоологи забыли записать количество кроликов в клетках их Всемирного заповедника. Но они знают, что было %2$s клеток и %3$s кроликов в каждой. Запишите ответ в той же системе счисления.</string>
+    <string name="Number3Task">В %1$s-чном мире случилась беда. Зоологи забыли записать количество кроликов в клетках их Всемирного заповедника. Но они знают, что было %2$s клеток и %3$s кроликов в каждой. Помогите им сосчитать кроликов. Запишите ответ в той же системе счисления.</string>
     <string name="editTextHint">Пишите здесь</string>>
 </resources>
